fix: encode
