+++ src/retrievals/models/base.py
@@ -1,14 +1,11 @@
 import logging
 from abc import ABC, abstractmethod
-from contextlib import nullcontext
 from typing import List, Optional, Tuple, Union
 
 import numpy as np
 import torch
 import torch.distributed as dist
 import torch.nn as nn
-from torch.utils.data import DataLoader
-from tqdm.auto import tqdm
 from transformers import PreTrainedTokenizer
 
 logger = logging.getLogger(__name__)
@@ -31,37 +28,6 @@ class Base(ABC, torch.nn.Module):
     def forward(self, *args, **kwargs):
         """Pytorch forward method."""
         raise NotImplementedError
-
-    def _encode_from_loader(
-        self,
-        loader: DataLoader,
-        convert_to_numpy: bool = True,
-        device: str = None,
-        normalize_embeddings: bool = False,
-        show_progress_bar: bool = None,
-        **kwargs,
-    ) -> Union[List[torch.Tensor], np.ndarray, torch.Tensor]:
-        """Encode for sentence embedding"""
-        device = device or self.device
-        self.model.eval()
-        self.model.to(device)
-
-        all_embeddings = []
-
-        for idx, inputs in enumerate(tqdm(loader, desc="Encoding", disable=not show_progress_bar)):
-            with torch.autocast(device_type=device) if self.use_fp16 else nullcontext():
-                with torch.no_grad():
-                    inputs_on_device = {k: v.to(device) for k, v in inputs.items()}
-                    embeddings = self.forward_from_loader(inputs_on_device)
-                    embeddings = embeddings.detach()
-                    if normalize_embeddings:
-                        embeddings = torch.nn.functional.normalize(embeddings, p=2, dim=1)
-                    all_embeddings.append(embeddings)
-        if convert_to_numpy:
-            all_embeddings = np.concatenate([emb.cpu().numpy() for emb in all_embeddings], axis=0)
-        else:
-            all_embeddings = torch.concat(all_embeddings)
-        return all_embeddings
 
     def save_pretrained(self, path: str, safe_serialization: bool = True):
         """
